Move record and constraint viewsets into their own files.
